# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for toy rental service",
        "title": "ToyRental API",
        "contact": {
            "name": "Indrawansyah",
            "email": "indra@dev.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/rental": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rental"
                ],
                "summary": "Mendapatkan semua data rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Rental"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rental"
                ],
                "summary": "Insert rental baru",
                "parameters": [
                    {
                        "description": "Rental",
                        "name": "rental",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateRentalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Rental"
                        }
                    }
                }
            }
        },
        "/rental/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rental"
                ],
                "summary": "Mendapatkan data rental berdasarkan id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rental ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Rental"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiCookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rental"
                ],
                "summary": "Perpanjang sewa rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Rental",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data perpanjangan rental",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ExtendRentalRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "Rental"
                ],
                "summary": "Hapus rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rental ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
        "/rental/{id}/return": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rental"
                ],
                "summary": "Pengembalian rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rental ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data pengembalian rental",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ReturnRentalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Rental"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CreateRentalItemRequest": {
            "type": "object",
            "properties": {
                "condition_before": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "toy_id": {
                    "type": "string"
                }
            }
        },
        "entity.CreateRentalRequest": {
            "type": "object",
            "properties": {
                "expected_return_date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CreateRentalItemRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "rental_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.ExtendRentalRequest": {
            "type": "object",
            "required": [
                "new_expected_return_date"
            ],
            "properties": {
                "new_expected_return_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "entity.Rental": {
            "type": "object",
            "properties": {
                "actual_return_date": {
                    "type": "string"
                },
                "damage_fee": {
                    "type": "number"
                },
                "expected_return_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "late_fee": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "rental_date": {
                    "type": "string"
                },
                "rental_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.RentalItem"
                    }
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_rental_price": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.RentalItem": {
            "type": "object",
            "properties": {
                "condition_after": {
                    "type": "string"
                },
                "condition_before": {
                    "type": "string"
                },
                "damage_description": {
                    "type": "string"
                },
                "damage_fee": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "price_per_unit": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "rental_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "toy": {
                    "$ref": "#/definitions/entity.Toy"
                },
                "toy_id": {
                    "type": "string"
                }
            }
        },
        "entity.ReturnRentalItemRequest": {
            "type": "object",
            "required": [
                "condition_after",
                "rental_item_id"
            ],
            "properties": {
                "condition_after": {
                    "type": "string"
                },
                "damage_description": {
                    "type": "string"
                },
                "rental_item_id": {
                    "type": "string"
                }
            }
        },
        "entity.ReturnRentalRequest": {
            "type": "object",
            "required": [
                "actual_return_date",
                "items"
            ],
            "properties": {
                "actual_return_date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ReturnRentalItemRequest"
                    }
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "entity.Toy": {
            "type": "object",
            "properties": {
                "age_recommendation": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ToyCategory"
                    }
                },
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ToyImage"
                    }
                },
                "is_available": {
                    "type": "boolean"
                },
                "late_fee_per_day": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "primary_image": {
                    "type": "string"
                },
                "rental_price": {
                    "type": "number"
                },
                "replacement_price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "entity.ToyCategory": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.ToyImage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                }
            }
        },
...
        "response.APISuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "metadata": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}